[MEM] Changed malloc to size_t to fix MEMmalloc warnings while fixing conversion warnings around the compiler

+++ src/libsac2c/global/memory.h
@@ -51,7 +51,7 @@ extern mallocphaseinfo_t phasetable[]; // needed by check_mem
  *
  *   Should be used instead of `malloc'.
  */
-extern void *_MEMmalloc (int size, const char *file, int line,
+extern void *_MEMmalloc (size_t size, const char *file, int line,
                          const char *func) FUN_ATTR_MALLOC;
 #define MEMmalloc(size) _MEMmalloc (size, __FILE__, __LINE__, __func__)
 #define MEMmallocAt(size, file, line) _MEMmalloc (size, file, line, __func__)
